BackEnd_WAWA/requirements.txt
# FastAPI: Python 기반의 현대적이고 빠른 웹 프레임워크. API 서버를 구축하는 데 사용됩니다.
fastapi
# Uvicorn: FastAPI와 같은 ASGI 애플리케이션을 실행하기 위한 고성능 서버입니다.
uvicorn[standard]
# Pydantic: 파이썬의 타입 힌트를 사용하여 데이터 유효성 검사, 설정 관리, 데이터 모델링을 수행하는 라이브러리로, FastAPI의 핵심 부품입니다.
pydantic

# Supabase-py: Supabase의 데이터베이스, 인증 등 다양한 기능과 상호작용하기 위한 공식 Python 클라이언트 라이브러리입니다.
supabase
# SQLAlchemy : 파이썬 코드와 데이터베이스를 연결해주는 ORM(Object Relational Mapper, 객체 관계 매핑) 라이브러리입니다.
sqlalchemy
# Alembic : 데이터베이스의 구조 변경 이력을 관리하고 적용하는 버전 관리 시스템입니다. SQLAlchemy를 위한 데이터베이스 마이그레이션 도구입니다.
alembic
# Psycopg2-binary : PostgreSQL 데이터베이스 드라이버입니다.
psycopg2-binary

# Python-dotenv: .env 파일에 정의된 환경 변수(예: DB 접속 정보)를 파이썬 스크립트로 쉽게 불러올 수 있게 해주는 라이브러리입니다.
python-dotenv

# Python-jose : JWT(JSON Web Token)의 생성, 서명, 검증을 담당하는 라이브러리입니다.
python-jose[cryptography]
# Passlib : 비밀번호 암호화(Hashing)하기 위한 라이브러리입니다. bcrypt 알고리즘을 함께 설치합니다.
passlib[bcrypt]

# Google OAuth : Google 로그인
authlib

python-dotenv
postgrest
fastapi-apscheduler
pyfcm
requests
email-validator
gotrue
numpy
pyinstrument

passlib[bcrypt]
bcrypt
googletrans-py